Add files via upload
The sample input and output has been stored in separate variables. The initial for loop traverses to the assigned variable "input" and storing it in string_input, The while condition executes until the variable "final_input" equals to 0. The variable mod calculates the resultant modulus of variable final_input with 10. Therefore in accordance with the index position it prints the required statement. 
Finally when the sample input is 10,300,5 the resultant output will be OneZero, ThreeZeroZero, Five